HT_21/requirements.txt
Django~=4.0.0
requests~=2.26.0
djangorestframework~=3.13.1

